More mapper work.  CPU now reads/writes from the mapper.  A bit better support for large ROM sizes.  Mapper1 still not working, but getting futher.

+++ SaltyLib/SaltyLib.Shared/NES/NESRom.cpp
@@ -94,7 +94,7 @@ uint8_t NESRom::GetMapperId() const
 }
 
 
-uint16_t NESRom::GetCbPrgRom() const
+uint32_t NESRom::GetCbPrgRom() const
 {
 	return m_header.CbPrgRomData();
 }
